requirements.txt
flask
flask-cors
tensorflow
numpy
Pillow
gunicorn
